+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-## [1.1.5]
+## [1.1.5] - 2020-05-01
 ### Added
  - new formatting option `typeCaseOverride` which works just like `keywordCaseOverride` but only for type tokens. 
 ### Fixed